third_party/WebKit/LayoutTests/svg/custom/path-arc-invalid-expected.txt

CONSOLE ERROR: line 6: Error: Invalid value for <path> attribute d="M0 0 A0 0 0 0 0 0 0 0 0 0"
 If this text is visible the malformed SVG path is handled correctly.
